requirements.txt
openai-whisper
pysrt
moviepy
fastapi
uvicorn
ffmpeg